Customer controlled created , started testing app

+++ src/main/java/com/example/orderservice/exception/OrderNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.orderservice.exception;
 
 public class OrderNotFoundException extends BaseServiceException {
-    public OrderNotFoundException(Integer orderId) {
+    public OrderNotFoundException(Long orderId) {
         super("Order with id " + orderId + " not found", "ORDER_NOT_FOUND");
     }
 }
